splitting css and making login changes
