+++ react-app/src/components/SurveyForm.jsx
@@ -11,11 +11,11 @@ export default function PollForm({ uniqueId }) {
   const { pushError } = useContext(MessageContext);
 
   const options = [
-    { value: "1", label: "Strongly Agree" },
-    { value: "2", label: "Weakly Agree" },
-    { value: "3", label: "Neutral" },
-    { value: "4", label: "Weakly Disagree" },
     { value: "5", label: "Strongly Disagree" },
+    { value: "4", label: "Weakly Disagree" },
+    { value: "3", label: "Neutral" },
+    { value: "2", label: "Weakly Agree" },
+    { value: "1", label: "Strongly Agree" },
   ];
 
   function handleOptionChange(event) {
