+++ react-app/src/components/nav/nav.css
@@ -1,0 +1,90 @@
+.nav {
+  color: #14213d;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  text-align: center;
+  position: relative;
+  z-index: 998;
+  top: 0;
+  left: 0;
+  border-bottom: 2px solid #b8bcc5;
+}
+
+.nav ul {
+  padding: 0;
+  margin: 0;
+  list-style: none;
+  display: flex;
+  gap: 1rem;
+  font-size: 1.5rem;
+  position: relative;
+  width: 100%;
+}
+
+.nav a {
+  color: #14213d;
+  padding: 1rem;
+  text-decoration: none;
+  border-bottom: none;
+  margin-right: 1rem;
+}
+
+.nav a:hover {
+  color: #ffa500;
+  transition: 0.2s ease-in;
+}
+
+.nav li {
+  list-style: none;
+}
+
+.nav li a {
+  font-size: 1.3rem;
+  font-weight: 700;
+  letter-spacing: -0.05em;
+}
+
+.nav-logo {
+  height: 7vh;
+  width: 10vw;
+  background-image: url("../../imgs/logo.png");
+  background-size: 14vh;
+  background-repeat: no-repeat;
+  z-index: 999;
+  background-size: contain;
+  background-position: center;
+}
+
+.site-title {
+  padding-left: 6vw;
+}
+
+@media screen and (max-width: 450px) {
+  html,
+  body {
+    margin: 0;
+    padding: 0;
+  }
+
+  .nav {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 0;
+    width: 100vw;
+  }
+
+  .nav li {
+    display: none;
+    padding: 0;
+  }
+
+  .site-title {
+    display: flex;
+    justify-content: center;
+    text-align: center;
+    width: 100%;
+  }
+}
