+++ react-app/src/imgs/front-page-comp.png
