+++ react-app/src/imgs/front-page-img.png
