+++ react-app/src/imgs/front-page-scientists.png
