+++ react-app/src/index.jsx
@@ -1,8 +1,7 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-
 import App from "./App";
-import "./App.css";
+import "./styles/global.css";
 
 const container = document.getElementById("root");
 const root = createRoot(container);
