+++ react-app/src/pages/about/About.jsx
@@ -1,6 +1,7 @@
 import React from "react";
-import NavBar from "../components/NavBar";
-import Footer from "../components/Footer";
+import NavBar from "../../components/nav/NavBar";
+import Footer from "../../components/footer/Footer";
+import "./about.css";
 
 export default function About() {
   return (
