+++ react-app/src/pages/dashboard/Dashboard.jsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect, useRef, useContext } from "react";
 import { Navigate } from "react-router-dom";
 import axios from "axios";
-import NavBar from "../components/NavBar";
-import Footer from "../components/Footer";
-import Table from "../components/DashboardTable";
-import CreateContainer from "../components/CreateContainer";
-import AddParticipants from "../components/AddParticipants";
-import { AuthContext } from "../components/AuthContext";
-import { MessageContext } from "../components/MessageHandler";
+import NavBar from "../../components/nav/NavBar";
+import Footer from "../../components/footer/Footer";
+import Table from "../../components/DashboardTable";
+import CreateContainer from "../../components/CreateContainer";
+import AddParticipants from "../../components/AddParticipants";
+import { AuthContext } from "../../components/AuthContext";
+import { MessageContext } from "../../components/MessageHandler";
 
 export default function Dashboard() {
   const [showCreateContainer, setShowCreateContainer] = useState(false);
