+++ react-app/src/pages/error/Error.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import NavBar from "../components/NavBar";
-import Footer from "../components/Footer";
+import NavBar from "../../components/nav/NavBar";
+import Footer from "../../components/footer/Footer";
 
 export default function Error() {
   return (
