+++ react-app/src/pages/login/login.css
@@ -1,0 +1,80 @@
+.login--container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.login-box {
+  height: 100vh;
+  width: 100vw;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.login--content {
+  width: 90%;
+  max-width: 500px;
+  height: auto;
+  padding: 2rem;
+  border-radius: 5px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  border: 2px solid rgba(51, 51, 51, 1);
+  background-color: #eeecec;
+}
+
+.login--form {
+  height: auto;
+  width: 90%;
+  max-width: 400px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+  margin: 1rem;
+}
+
+.login--form div {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.login--form label {
+  font-size: 1rem;
+  font-weight: 500;
+  padding: 0.5rem;
+  text-align: center;
+}
+
+.login--form input {
+  font-size: 1rem;
+  width: 100%;
+  padding: 0.5rem;
+  margin-top: 1rem;
+  text-align: left;
+}
+
+.login--form button {
+  font-size: 1rem;
+  font-weight: 500;
+}
+
+.login--pic {
+  width: 100%;
+  height: 10vw;
+  background-position: center;
+  background-image: url("../../imgs/front-page-img.png");
+  background-repeat: no-repeat;
+  background-size: contain;
+}
+
+.login--form h1 {
+  font-size: 1rem;
+}
