+++ react-app/src/pages/survey/survey.css
@@ -1,0 +1,255 @@
+.poll--total {
+  position: relative;
+  width: 100vw;
+  min-height: 100vh;
+  backdrop-filter: blur(0px);
+}
+
+.background-blur {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  background-image: url("../../imgs/background.png");
+  background-position: bottom right;
+  background-repeat: no-repeat;
+  background-size: 50% 80%;
+  filter: blur(25px) opacity(0.15);
+  z-index: -1;
+}
+
+.mirror {
+  transform: scaleX(-1);
+}
+
+.poll {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  width: 100%;
+  height: 100vh;
+  flex-grow: 1;
+  padding: 1rem;
+}
+
+.poll--box {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 700px;
+  height: 600px;
+  border-radius: 5px;
+  border: 2px solid rgba(51, 51, 51, 1);
+  background-color: #eeecec;
+}
+
+.poll--blurb {
+  font-weight: 300;
+  font-size: 1.3rem;
+  letter-spacing: -0.05em;
+  color: #5c667d;
+  align-items: center;
+  margin-top: 2rem;
+  margin-left: 2rem;
+  margin-right: 2rem;
+  text-align: center;
+}
+
+.poll--question {
+  font-weight: 600;
+  font-size: 1.8rem;
+  color: #14213d;
+  margin-top: 1rem;
+  align-items: center;
+  margin-bottom: 1rem;
+  margin-left: 2rem;
+  margin-right: 2rem;
+  text-align: center;
+}
+
+.poll--options-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding-top: 1rem;
+}
+
+.poll--options {
+  list-style: none;
+  font-weight: 500;
+  font-size: 1.5rem;
+  letter-spacing: -0.05em;
+  color: #14213d;
+  padding: 0;
+  margin: 0;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+}
+
+.poll--options li {
+  display: flex;
+  flex-direction: column;
+  padding: 2rem;
+}
+
+.poll--submit {
+  cursor: pointer;
+  border: 2px solid rgba(51, 51, 51, 0.7);
+  font-size: 1.3rem;
+  background-color: rgba(51, 51, 51, 0.7);
+  margin-bottom: 2rem;
+}
+
+.poll--submit-active:hover {
+  opacity: 100%;
+}
+
+.poll--submit-active {
+  transition: 0.2s ease-in;
+  background-color: #14213d;
+  color: #ffffff;
+}
+
+.poll--submit-active:hover {
+  opacity: 70%;
+  transition: 0.2s ease-in-out;
+}
+
+/* poll form */
+
+.poll--box form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.poll--options {
+  padding: 0rem 2rem 0rem 2rem;
+}
+
+.poll--options li {
+  display: flex;
+  align-items: center;
+  margin-bottom: 1rem;
+}
+
+.poll--checkbox {
+  appearance: none;
+  -webkit-appearance: none;
+  position: relative;
+  width: 1.5rem;
+  height: 1.5rem;
+  margin-right: 1rem;
+  outline: none;
+  border: 2px solid #14213d;
+  border-radius: 5px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.poll--checkbox + label {
+  transition-duration: 0.2s;
+}
+
+.poll--checkbox + label {
+  font-size: 1.3rem;
+}
+
+.poll--checkbox:checked {
+  content: "\2715";
+  font-size: 1.5rem;
+  color: #14213d;
+}
+
+.poll--checkbox:checked + label:before {
+  position: absolute;
+  top: 3px;
+  left: 3px;
+  width: 10px;
+  height: 10px;
+  background-color: #14213d;
+  border-radius: 2px;
+}
+
+.poll--checkbox:checked + label:before {
+  display: none;
+}
+
+.poll--checkbox:checked:after {
+  content: "\2715";
+  font-size: 1.5rem;
+  color: #14213d;
+}
+
+.poll--checkbox:not(:checked):hover {
+  content: "\2715";
+  font-size: 1.5rem;
+  color: #5c667d;
+  transition: 0.1s ease-in-out;
+}
+
+.poll--checkbox:not(:checked):hover::after {
+  content: "\2715";
+  font-size: 1.5rem;
+  color: #5c667d;
+  transition: 0.1s ease-in-out;
+}
+
+.poll--option-spacer label {
+  /* padding-top: 1.5rem !important;
+      padding-bottom: 1.5rem !important; */
+  font-size: 1rem;
+  color: #888e9b;
+}
+.poll--option-spacer input {
+  /* padding-top: 1.5rem !important;
+      padding-bottom: 1.5rem !important;  */
+  border-color: #888e9b !important;
+  height: 1rem;
+  width: 1rem;
+}
+
+.poll--option-spacer {
+  /* padding-top: 1.5rem !important;
+      padding-bottom: 1.5rem !important; */
+  padding-left: 5rem !important;
+}
+
+@media screen and (max-width: 450px) {
+  .poll--box {
+    width: 95%;
+    margin-top: 1rem;
+    height: none;
+  }
+  .poll {
+    padding: 0rem;
+    margin-left: 0rem;
+    margin-right: 0rem;
+    border: #d9d1d3;
+  }
+  .poll--total {
+    height: auto;
+  }
+
+  .background-blur {
+    height: 0%;
+  }
+  .poll--options {
+    font-size: 1.2rem;
+  }
+  .poll--blurb {
+    font-size: 1.2rem;
+  }
+  .poll--question {
+    font-size: 1.5rem;
+  }
+  .poll--submit {
+    font-size: 1.2rem;
+  }
+}
