+++ react-app/src/styles/global.css
@@ -1,0 +1,197 @@
+@import url("https://fonts.googleapis.com/css2?family=Inter:wght@400;600;700;800&display=swap");
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+html,
+body {
+  height: 100vh;
+  width: 100vw;
+  margin: 0;
+  padding: 0;
+  font-family: "Inter", sans-serif;
+  background-color: #f5f5f5;
+  color: #14213d;
+  letter-spacing: -0.05em;
+}
+
+.div {
+  width: 100vw;
+  height: 100vh;
+}
+
+a {
+  color: #3c096c;
+}
+
+h1,
+h2,
+h3,
+p,
+ol,
+ul {
+  font-family: "Inter";
+  font-style: normal;
+}
+
+h1 {
+  font-size: 4rem;
+  font-weight: 700;
+  padding: 1rem 0rem 1rem 0rem;
+}
+
+h2 {
+  font-size: 3rem;
+  font-weight: 700;
+  padding: 1rem 0rem 1rem 0rem;
+}
+
+h3 {
+  font-size: 2rem;
+  font-weight: 700;
+  padding: 1rem 0rem 1rem 0rem;
+}
+
+p,
+ol,
+ul {
+  font-size: 1.2rem;
+  font-weight: 400;
+}
+
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  height: 100vh;
+  width: 100vw;
+}
+
+.centered {
+  justify-content: center;
+}
+
+.padding {
+  padding: 2rem;
+}
+
+.button {
+  padding: 8px 16px;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 1.2rem;
+  border: none;
+  width: 20rem;
+  text-align: center;
+  border: 2px solid #007bff;
+  font-weight: 500;
+  margin-top: 1rem;
+  font-size: 1.5rem;
+  background-color: #007bff;
+  color: #ffffff;
+}
+
+.button:hover {
+  filter: opacity(0.7);
+  transition: 0.2s ease-in-out;
+}
+
+@media (max-width: 450px) {
+  h1 {
+    font-size: 2.5rem;
+  }
+
+  h2 {
+    font-size: 2rem;
+  }
+
+  h3,
+  p,
+  ol,
+  ul {
+    font-size: 1rem;
+  }
+}
+
+.alert {
+  border: 1px solid #000;
+  border-radius: 4px;
+  background: #eee;
+  display: block;
+  float: left;
+  width: 50%;
+  padding: 8px;
+  margin: 8px 8px 8px 25%;
+  position: relative;
+  z-index: 9999;
+  filter: drop-shadow(2px 4px 4px rgba(0, 0, 0, 0.25));
+}
+
+@media only screen and (max-width: 600px) {
+  .alert {
+    margin: 8px;
+    width: 97%;
+  }
+}
+
+.alert--float {
+  position: absolute;
+  width: 100%;
+  margin-top: 10px;
+}
+
+.alert.alert--error {
+  border: 1px solid #a22;
+  background: #faa;
+}
+
+.alert.alert--warning {
+  border: 1px solid #ffa500;
+  background: #fd5;
+}
+
+.alert.alert--info {
+  border: 1px solid #007bff;
+  background: #6bbcff;
+}
+
+.alert > .alert--symbol {
+  max-width: 10px;
+  float: left;
+  padding: 8px;
+}
+
+.alert > .alert--dismiss {
+  float: right;
+  margin-right: 16px;
+  cursor: pointer;
+}
+
+.alert > .alert--body {
+  margin-left: 50px;
+  margin-right: 50px;
+}
+
+.alert > span.alert--title {
+}
+
+.alert > span.alert--message {
+}
+
+.animate-entry {
+  animation: animate-entry 0.5s both;
+}
+@keyframes animate-entry {
+  0% {
+    transform: translateY(-10px);
+    opacity: 0;
+  }
+  100% {
+    transform: translateY(0);
+    opacity: 1;
+  }
+}
